Extract "exception" assertion to own file
