+++ lib/referee.js
@@ -2,7 +2,6 @@
 
 var actualAndExpectedMessageValues = require("./actual-and-expected-message-values");
 var assertArgNum = require("./assert-arg-num");
-var captureException = require("./capture-exception");
 var interpolatePosArg = require("./interpolate-pos-arg");
 var interpolateProperties = require("./interpolate-properties");
 var expect = require("./expect");
@@ -202,70 +201,7 @@ require("./assertions/is-nan")(referee);
 require("./assertions/is-array")(referee);
 require("./assertions/is-array-like")(referee);
 require("./assertions/keys")(referee);
-
-referee.captureException = captureException;
-
-referee.add("exception", {
-    assert: function (callback) {
-        var matcher = arguments[1];
-        var customMessage = arguments[2];
-
-        if (typeof matcher === "string") {
-            customMessage = matcher;
-            matcher = undefined;
-        }
-
-        this.expected = matcher;
-        this.customMessage = customMessage;
-
-        var err = captureException(callback);
-
-        if (err) {
-            this.actualExceptionType = err.name;
-            this.actualExceptionMessage = err.message;
-            this.actualExceptionStack = err.stack;
-        }
-
-        if (!err) {
-            if (typeof matcher === "object") {
-                return this.fail("typeNoExceptionMessage");
-            }
-            return this.fail("message");
-
-        }
-
-        if (typeof matcher === "object" && !referee.match(err, matcher)) {
-            return this.fail("typeFailMessage");
-        }
-
-        if (typeof matcher === "function" && matcher(err) !== true) {
-            return this.fail("matchFailMessage");
-        }
-
-        return true;
-    },
-
-    refute: function (callback) {
-        var err = captureException(callback);
-
-        if (err) {
-            this.customMessage = arguments[1];
-            this.actualExceptionType = err.name;
-            this.actualExceptionMessage = err.message;
-            return false;
-        }
-
-        return true;
-    },
-
-    expectation: "toThrow",
-    assertMessage: "${customMessage}Expected exception",
-    refuteMessage: "${customMessage}Expected not to throw but threw ${actualExceptionType} (${actualExceptionMessage})"
-});
-
-referee.assert.exception.typeNoExceptionMessage = "${customMessage}Expected ${expected} but no exception was thrown";
-referee.assert.exception.typeFailMessage = "${customMessage}Expected ${expected} but threw ${actualExceptionType} (${actualExceptionMessage})\n${actualExceptionStack}";
-referee.assert.exception.matchFailMessage = "${customMessage}Expected thrown ${actualExceptionType} (${actualExceptionMessage}) to pass matcher function";
+require("./assertions/exception")(referee);
 
 referee.add("near", {
     assert: function (actual, expected, delta) {
